Restructure stdlib tests to match 1:1 with implementation files
Split monolithic test files into individual files that directly correspond to stdlib implementation files. This provides:

**Benefits:**
- 1:1 mapping between implementation and test files (e.g., `src/javascript/stdlib/array/push.ts` → `tests/javascript/stdlib/array/push.test.ts`)
- Easier navigation and test discovery
- Faster test iteration (run only relevant tests)
- Better test failure messages (specific to method)
- Parallel test execution by Vitest
- Clearer organization with `_` prefix for shared/meta tests

**Structure created:**
```
tests/
├── javascript/stdlib/
│   ├── array/          # 20 method tests + 2 shared tests
│   │   ├── length.test.ts, at.test.ts, push.test.ts, pop.test.ts, etc.
│   │   └── _computed-access.test.ts (shared error tests)
│   └── string/         # 2 method tests + 3 meta tests
│       ├── length.test.ts, toLowerCase.test.ts, toUpperCase.test.ts
│       ├── _computed-access.test.ts
│       ├── _language-features.test.ts
│       └── _unimplemented.test.ts
└── python/stdlib/
    ├── list/           # 2 method tests + 1 shared test
    │   ├── __len__.test.ts
    │   └── _chaining-and-access.test.ts
    └── string/         # 2 method tests + 3 meta tests
        ├── upper.test.ts, lower.test.ts
        ├── _chaining.test.ts
        ├── _attribute-access.test.ts
        └── _immutability.test.ts
```

**Changes:**
- Created 33 new focused test files
- Deleted 4 monolithic test files:
  - `tests/javascript/array-properties-methods.test.ts` (1143 lines)
  - `tests/javascript/string-properties-methods.test.ts` (403 lines)
  - `tests/python/list-attributes.test.ts` (123 lines)
  - `tests/python/string-methods.test.ts` (266 lines)
- Net: +29 files, but significantly better organization

All 2624 tests continue to pass.
